adds foods.js controller file, renames foods.js model to food.js and alters files to ensure passing tests; preparing for a foods controller that uses the model for posting
