+++ lib/controllers/foods.js
@@ -1,0 +1,1 @@
+const food = require('../models/food.js')
