+++ test/foods_test.js
@@ -5,7 +5,7 @@ const pry = require('pryjs')
 const environment = process.env.NODE_ENV || 'test';
 const configuration = require('../knexfile')[environment];
 const database = require('knex')(configuration);
-const foods = require('../lib/models/foods')
+const foods = require('../lib/models/food')
 
 describe('Foods', function(){
   before(function(done){
